🚚rename[#50]: 공통 Exception 패키지 변경
