+++ src/test/java/com/sillim/recordit/calendar/domain/vo/CalendarColorHexTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.junit.jupiter.api.Assertions.*;
 
 import com.sillim.recordit.global.exception.ErrorCode;
-import com.sillim.recordit.global.exception.schedule.InvalidColorHexException;
+import com.sillim.recordit.global.exception.common.InvalidColorHexException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
